aisearchservice/requirements.txt
PyJWT
streamlit
fastapi[standard]
python-multipart
uvicorn
langchain-community
opensearch-py
pdf2image
sentence_transformers
accelerate
unstructured
unstructured[all-docs]
chromadb
yandex-cloud-ml-sdk